Fix a problematic migration

+++ db/migrate/20140425160626_change_type_of_assessment_question_option_value.rb
@@ -1,5 +1,5 @@
 class ChangeTypeOfAssessmentQuestionOptionValue < ActiveRecord::Migration
   def change
-    change_column :assessment_question_options, :value, :float
+    change_column :assessment_question_options, :value, :decimal
   end
 end
